# backend/oasgen/src/main/java/com/codevision/codevisionbackend/api/generated/SecurityApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.codevision.codevisionbackend.api.generated;

import com.codevision.codevisionbackend.api.model.ProjectLoggerInsightsResponse;
import com.codevision.codevisionbackend.api.model.ProjectPiiPciResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-06T21:06:48.004750-05:00[America/New_York]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Security", description = "Security and compliance visibility endpoints.")
public interface SecurityApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /project/{projectId}/export/logs.csv : Download logger insights as CSV
     *
     * @param projectId  (required)
     * @return CSV export generated. (status code 200)
     *         or No logger insights were found to export. (status code 404)
     */
    @Operation(
        operationId = "exportProjectLogsCsv",
        summary = "Download logger insights as CSV",
        tags = { "Security" },
        responses = {
            @ApiResponse(responseCode = "200", description = "CSV export generated.", content = {
                @Content(mediaType = "text/csv", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            }),
            @ApiResponse(responseCode = "404", description = "No logger insights were found to export.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/project/{projectId}/export/logs.csv",
        produces = { "text/csv" }
    )
    
    default ResponseEntity<org.springframework.core.io.Resource> exportProjectLogsCsv(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /project/{projectId}/export/logs.pdf : Download logger insights as PDF
     *
     * @param projectId  (required)
     * @return PDF export generated. (status code 200)
     *         or No logger insights were found to export. (status code 404)
     */
    @Operation(
        operationId = "exportProjectLogsPdf",
        summary = "Download logger insights as PDF",
        tags = { "Security" },
        responses = {
            @ApiResponse(responseCode = "200", description = "PDF export generated.", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            }),
            @ApiResponse(responseCode = "404", description = "No logger insights were found to export.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/project/{projectId}/export/logs.pdf",
        produces = { "application/pdf" }
    )
    
    default ResponseEntity<org.springframework.core.io.Resource> exportProjectLogsPdf(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /project/{projectId}/export/pii.csv : Download PCI/PII findings as CSV
     *
     * @param projectId  (required)
     * @return CSV export generated. (status code 200)
     *         or No PCI/PII findings were found to export. (status code 404)
     */
    @Operation(
        operationId = "exportProjectPiiCsv",
        summary = "Download PCI/PII findings as CSV",
        tags = { "Security" },
        responses = {
            @ApiResponse(responseCode = "200", description = "CSV export generated.", content = {
                @Content(mediaType = "text/csv", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            }),
            @ApiResponse(responseCode = "404", description = "No PCI/PII findings were found to export.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/project/{projectId}/export/pii.csv",
        produces = { "text/csv" }
    )
    
    default ResponseEntity<org.springframework.core.io.Resource> exportProjectPiiCsv(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /project/{projectId}/export/pii.pdf : Download PCI/PII findings as PDF
     *
     * @param projectId  (required)
     * @return PDF export generated. (status code 200)
     *         or No PCI/PII findings were found to export. (status code 404)
     */
    @Operation(
        operationId = "exportProjectPiiPdf",
        summary = "Download PCI/PII findings as PDF",
        tags = { "Security" },
        responses = {
            @ApiResponse(responseCode = "200", description = "PDF export generated.", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            }),
            @ApiResponse(responseCode = "404", description = "No PCI/PII findings were found to export.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/project/{projectId}/export/pii.pdf",
        produces = { "application/pdf" }
    )
    
    default ResponseEntity<org.springframework.core.io.Resource> exportProjectPiiPdf(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /project/{projectId}/logger-insights : Retrieve logger insight catalog
     * Returns all log statements discovered in the repository, annotated with PII/PCI risk flags.
     *
     * @param projectId Unique identifier of the project. (required)
     * @return Logger insights retrieved successfully. (status code 200)
     *         or No logger insights found for the supplied identifier. (status code 404)
     */
    @Operation(
        operationId = "getProjectLoggerInsights",
        summary = "Retrieve logger insight catalog",
        description = "Returns all log statements discovered in the repository, annotated with PII/PCI risk flags.",
        tags = { "Security" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Logger insights retrieved successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProjectLoggerInsightsResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "No logger insights found for the supplied identifier.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/project/{projectId}/logger-insights",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ProjectLoggerInsightsResponse> getProjectLoggerInsights(
        @Parameter(name = "projectId", description = "Unique identifier of the project.", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"loggerInsights\" : [ { \"variables\" : [ \"variables\", \"variables\" ], \"logLevel\" : \"logLevel\", \"filePath\" : \"filePath\", \"className\" : \"className\", \"piiRisk\" : true, \"lineNumber\" : 1, \"pciRisk\" : true, \"messageTemplate\" : \"messageTemplate\" }, { \"variables\" : [ \"variables\", \"variables\" ], \"logLevel\" : \"logLevel\", \"filePath\" : \"filePath\", \"className\" : \"className\", \"piiRisk\" : true, \"lineNumber\" : 1, \"pciRisk\" : true, \"messageTemplate\" : \"messageTemplate\" } ], \"projectId\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /project/{projectId}/pii-pci : Retrieve PCI/PII scan results
     * Returns potential sensitive data matches discovered across the repository.
     *
     * @param projectId Unique identifier of the project. (required)
     * @return PCI/PII findings retrieved successfully. (status code 200)
     *         or No scan results were found for the supplied identifier. (status code 404)
     */
    @Operation(
        operationId = "getProjectPiiPciFindings",
        summary = "Retrieve PCI/PII scan results",
        description = "Returns potential sensitive data matches discovered across the repository.",
        tags = { "Security" },
        responses = {
            @ApiResponse(responseCode = "200", description = "PCI/PII findings retrieved successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProjectPiiPciResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "No scan results were found for the supplied identifier.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/project/{projectId}/pii-pci",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ProjectPiiPciResponse> getProjectPiiPciFindings(
        @Parameter(name = "projectId", description = "Unique identifier of the project.", required = true, in = ParameterIn.PATH) @PathVariable("projectId") Long projectId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"findings\" : [ { \"snippet\" : \"snippet\", \"severity\" : \"severity\", \"ignored\" : true, \"matchType\" : \"matchType\", \"filePath\" : \"filePath\", \"lineNumber\" : 5 }, { \"snippet\" : \"snippet\", \"severity\" : \"severity\", \"ignored\" : true, \"matchType\" : \"matchType\", \"filePath\" : \"filePath\", \"lineNumber\" : 5 } ], \"projectId\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}