# backend/oasgen/src/main/resources/oas/openapi.yaml
openapi: 3.0.3
info:
  title: CodeVision API
  version: 0.3.0
  description: |
    API for initiating repository analysis and retrieving project overview metadata
    produced by the CodeVision ingestion backend.
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Analysis
    description: Endpoints used to trigger repository analysis workflows.
  - name: Project
    description: Endpoints that expose stored project overview information.
  - name: Security
    description: Security and compliance visibility endpoints.
paths:
  /analyze:
    post:
      operationId: analyzeRepository
      tags:
        - Analysis
      summary: Trigger analysis for a repository
      description: Starts a background analysis workflow for the supplied repository URL.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./components/schemas.yaml#/AnalyzeRequest
            examples:
              analyzeRequest:
                $ref: ./examples/analyze-post-request.json
      responses:
        '200':
          description: Analysis request accepted and the snapshot metadata has been persisted.
          content:
            application/json:
              schema:
                $ref: ./components/schemas.yaml#/AnalyzeResponse
              examples:
                analyzeResponse:
                  $ref: ./examples/analyze-post-response.json
        '400':
          description: The request payload failed validation.
        '401':
          description: Missing or invalid API key.
  /project/{projectId}/overview:
    get:
      operationId: getProjectOverview
      tags:
        - Project
      summary: Retrieve project overview snapshot
      description: Returns the most recently stored overview snapshot for the selected project.
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier of the project.
          schema:
            type: integer
            format: int64
          examples:
            projectId:
              $ref: ./examples/project-overview-request.json
      responses:
        '200':
          description: Snapshot metadata was found.
          content:
            application/json:
              schema:
                $ref: ./components/schemas.yaml#/ParsedDataResponse
              examples:
                projectOverview:
                  $ref: ./examples/project-overview-response.json
        '404':
          description: No overview snapshot is available for the supplied identifier.
  /project/{projectId}/api-endpoints:
    get:
      operationId: getProjectApiEndpoints
      tags:
        - Project
      summary: Retrieve API endpoint catalog
      description: Returns all API endpoints discovered for the selected project grouped by protocol.
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier of the project.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Endpoint catalog retrieved successfully.
          content:
            application/json:
              schema:
                $ref: ./components/schemas.yaml#/ProjectApiEndpointsResponse
        '404':
          description: No catalog is available for the supplied identifier.
  /project/{projectId}/db-analysis:
    get:
      operationId: getProjectDbAnalysis
      tags:
        - Project
      summary: Retrieve database analysis details
      description: Returns entity metadata and repository/DAO operations captured during analysis.
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier of the project.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Database analysis retrieved successfully.
          content:
            application/json:
              schema:
                $ref: ./components/schemas.yaml#/ProjectDbAnalysisResponse
        '404':
          description: No database analysis is available for the supplied identifier.
  /project/{projectId}/logger-insights:
    get:
      operationId: getProjectLoggerInsights
      tags:
        - Security
      summary: Retrieve logger insight catalog
      description: Returns all log statements discovered in the repository, annotated with PII/PCI risk flags.
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier of the project.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Logger insights retrieved successfully.
          content:
            application/json:
              schema:
                $ref: ./components/schemas.yaml#/ProjectLoggerInsightsResponse
        '404':
          description: No logger insights found for the supplied identifier.
  /project/{projectId}/pii-pci:
    get:
      operationId: getProjectPiiPciFindings
      tags:
        - Security
      summary: Retrieve PCI/PII scan results
      description: Returns potential sensitive data matches discovered across the repository.
      parameters:
        - name: projectId
          in: path
          required: true
          description: Unique identifier of the project.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: PCI/PII findings retrieved successfully.
          content:
            application/json:
              schema:
                $ref: ./components/schemas.yaml#/ProjectPiiPciResponse
        '404':
          description: No scan results were found for the supplied identifier.
  /project/{projectId}/export/logs.csv:
    get:
      operationId: exportProjectLogsCsv
      tags:
        - Security
      summary: Download logger insights as CSV
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: CSV export generated.
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '404':
          description: No logger insights were found to export.
  /project/{projectId}/export/logs.pdf:
    get:
      operationId: exportProjectLogsPdf
      tags:
        - Security
      summary: Download logger insights as PDF
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: PDF export generated.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: No logger insights were found to export.
  /project/{projectId}/export/pii.csv:
    get:
      operationId: exportProjectPiiCsv
      tags:
        - Security
      summary: Download PCI/PII findings as CSV
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: CSV export generated.
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '404':
          description: No PCI/PII findings were found to export.
  /project/{projectId}/export/pii.pdf:
    get:
      operationId: exportProjectPiiPdf
      tags:
        - Security
      summary: Download PCI/PII findings as PDF
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: PDF export generated.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: No PCI/PII findings were found to export.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    AnalyzeRequest:
      $ref: ./components/schemas.yaml#/AnalyzeRequest
    AnalyzeResponse:
      $ref: ./components/schemas.yaml#/AnalyzeResponse
    BuildInfo:
      $ref: ./components/schemas.yaml#/BuildInfo
    ClassMetadataSummary:
      $ref: ./components/schemas.yaml#/ClassMetadataSummary
    MetadataDump:
      $ref: ./components/schemas.yaml#/MetadataDump
    OpenApiSpec:
      $ref: ./components/schemas.yaml#/OpenApiSpec
    ParsedDataResponse:
      $ref: ./components/schemas.yaml#/ParsedDataResponse
    SpecDocument:
      $ref: ./components/schemas.yaml#/SpecDocument
    SoapServiceSummary:
      $ref: ./components/schemas.yaml#/SoapServiceSummary
    SoapPortSummary:
      $ref: ./components/schemas.yaml#/SoapPortSummary
    ApiEndpoint:
      $ref: ./components/schemas.yaml#/ApiEndpoint
    ApiSpecArtifact:
      $ref: ./components/schemas.yaml#/ApiSpecArtifact
    AssetInventory:
      $ref: ./components/schemas.yaml#/AssetInventory
    ImageAsset:
      $ref: ./components/schemas.yaml#/ImageAsset
    ProjectApiEndpointsResponse:
      $ref: ./components/schemas.yaml#/ProjectApiEndpointsResponse
    DbAnalysis:
      $ref: ./components/schemas.yaml#/DbAnalysis
    DbEntitySummary:
      $ref: ./components/schemas.yaml#/DbEntitySummary
    DbFieldSummary:
      $ref: ./components/schemas.yaml#/DbFieldSummary
    DbRelationshipSummary:
      $ref: ./components/schemas.yaml#/DbRelationshipSummary
    DaoOperationDetails:
      $ref: ./components/schemas.yaml#/DaoOperationDetails
    ProjectDbAnalysisResponse:
      $ref: ./components/schemas.yaml#/ProjectDbAnalysisResponse